Refactored the whole library to have more granular keys and a VFS.
Now users can use any key types they want!
