+++ tests/lib.rs
@@ -4,12 +4,12 @@ use std::error::Error;
 use std::fmt;
 use std::fs::File;
 use std::io::{Read, Write};
-use warmy::{Key, Load, LogicalKey, Loaded, PathKey, Storage};
+use warmy::{Load, FSKey, LogicalKey, Loaded, Res, Storage};
 
 mod utils;
 
 /// Timeout in milliseconds to wait before determining that there’s something wrong with notify.
-const QUEUE_TIMEOUT_MS: u64 = 10000; // 10s
+const QUEUE_TIMEOUT_MS: u64 = 5000; // 5s
 
 #[derive(Debug, Eq, PartialEq)]
 struct Foo(String);
@@ -30,7 +30,7 @@ impl fmt::Display for FooErr {
 }
 
 impl Load for Foo {
-  type Key = PathKey;
+  type Key = FSKey;
 
   type Error = FooErr;
 
@@ -69,7 +69,7 @@ impl fmt::Display for BarErr {
 }
 
 impl Load for Bar {
-  type Key = PathKey;
+  type Key = FSKey;
 
   type Error = BarErr;
 
@@ -134,13 +134,13 @@ impl Load for LogicalFoo {
   type Error = LogicalFooErr;
 
   fn load(key: Self::Key, storage: &mut Storage) -> Result<Loaded<Self>, Self::Error> {
-    let key: Key<Foo> = Key::path(key.as_str()).expect("logical foo path");
-    let foo = storage.get(&key).unwrap();
+    let fs_key = FSKey::new(key.as_str());
+    let foo: Res<Foo> = storage.get(&fs_key).unwrap();
 
     let content = foo.borrow().0.clone();
     let zoo = LogicalFoo(content);
 
-    let r = Loaded::with_deps(zoo, vec![key.into()]);
+    let r = Loaded::with_deps(zoo, vec![fs_key.into()]);
     Ok(r)
   }
 }
@@ -155,6 +155,8 @@ fn foo() {
   utils::with_store(|mut store| {
     let expected1 = "Hello, world!".to_owned();
     let expected2 = "Bye!".to_owned();
+
+    let key = FSKey::new("foo.txt");
     let path = store.root().join("foo.txt");
 
     {
@@ -162,9 +164,45 @@ fn foo() {
       let _ = fh.write_all(expected1.as_bytes());
     }
 
-    let key: Key<Foo> = Key::path(&path).unwrap();
+    let r: Res<Foo> = store.get(&key).expect("object should be present at the given key");
 
-    let r = store.get(&key).expect("object should be present at the given key");
+    assert_eq!(r.borrow().0, expected1);
+
+    {
+      let mut fh = File::create(&path).unwrap();
+      let _ = fh.write_all(expected2.as_bytes());
+    }
+
+    let start_time = ::std::time::Instant::now();
+    loop {
+      store.sync();
+
+      if r.borrow().0.as_str() == expected2.as_str() {
+        break;
+      }
+
+      if start_time.elapsed() >= ::std::time::Duration::from_millis(QUEUE_TIMEOUT_MS) {
+        panic!("more than {} milliseconds were spent waiting for a filesystem event", QUEUE_TIMEOUT_MS);
+      }
+    }
+  })
+}
+
+#[test]
+fn foo_with_leading_slash() {
+  utils::with_store(|mut store| {
+    let expected1 = "Hello, world!".to_owned();
+    let expected2 = "Bye!".to_owned();
+
+    let key = FSKey::new("/foo.txt");
+    let path = store.root().join("foo.txt");
+
+    {
+      let mut fh = File::create(&path).unwrap();
+      let _ = fh.write_all(expected1.as_bytes());
+    }
+
+    let r: Res<Foo> = store.get(&key).expect("object should be present at the given key");
 
     assert_eq!(r.borrow().0, expected1);
 
@@ -191,6 +229,8 @@ fn foo() {
 #[test]
 fn two_same_paths_diff_types() {
   utils::with_store(|mut store| {
+    let foo_key = FSKey::new("a.txt");
+    let bar_key = foo_key.clone();
     let path = store.root().join("a.txt");
 
     // create a.txt
@@ -199,13 +239,10 @@ fn two_same_paths_diff_types() {
       let _ = fh.write_all(&b"foobarzoo"[..]);
     }
 
-    let foo_key: Key<Foo> = Key::path(&path).unwrap();
-    let bar_key: Key<Bar> = Key::path(&path).unwrap();
-
-    let foo = store.get(&foo_key).unwrap();
+    let foo: Res<Foo> = store.get(&foo_key).unwrap();
     assert_eq!(foo.borrow().0.as_str(), "foobarzoo");
 
-    let bar = store.get(&bar_key);
+    let bar: Result<Res<Bar>, _> = store.get(&bar_key);
     assert!(bar.is_err());
   })
 }
@@ -213,8 +250,8 @@ fn two_same_paths_diff_types() {
 #[test]
 fn logical_resource() {
   utils::with_store(|mut store| {
-    let key: Key<Zoo> = Key::logical("mem/uid/32197");
-    let zoo = store.get(&key).unwrap();
+    let key = LogicalKey::new("mem/uid/32197");
+    let zoo: Res<Zoo> = store.get(&key).unwrap();
     assert_eq!(zoo.borrow().0.as_str(), "mem/uid/32197");
   })
 }
@@ -224,6 +261,8 @@ fn logical_with_deps() {
   utils::with_store(|mut store| {
     let expected1 = "Hello, world!".to_owned();
     let expected2 = "Bye!".to_owned();
+
+    let foo_key = FSKey::new("foo.txt");
     let path = store.root().join("foo.txt");
 
     {
@@ -231,12 +270,10 @@ fn logical_with_deps() {
       let _ = fh.write_all(expected1.as_bytes());
     }
 
-    let foo_key: Key<Foo> = Key::path(&path).unwrap();
+    let _: Res<Foo> = store.get(&foo_key).expect("object should be present at the given key");
 
-    let _ = store.get(&foo_key).expect("object should be present at the given key");
-
-    let log_foo_key: Key<LogicalFoo> = Key::logical(path.to_str().unwrap());
-    let log_foo = store.get(&log_foo_key).unwrap();
+    let log_foo_key = LogicalKey::new(foo_key.as_path().to_str().unwrap());
+    let log_foo: Res<LogicalFoo> = store.get(&log_foo_key).unwrap();
 
     assert_eq!(log_foo.borrow().0.as_str(), "Hello, world!");
 
@@ -254,6 +291,7 @@ fn logical_with_deps() {
       }
 
       if start_time.elapsed() >= ::std::time::Duration::from_millis(QUEUE_TIMEOUT_MS) {
+        eprintln!("log_foo = {:?}", log_foo.borrow().0.as_str());
         panic!("more than {} milliseconds were spent waiting for a filesystem event", QUEUE_TIMEOUT_MS);
       }
     }
